Add *Of properties to RefResolver
The RefResolver currently skips any part of the schema that contains
'allOf', 'anyOf' or 'oneOf'. This prevents proper resolving of
references ($ref) that are inside of an array that is part of the above
mentioned properties.
Add the three property names to the array that is used to execute
resolveArrayOfSchemas().