Fix URI generic syntax delimit path components by slash ("`/`")
Parts of justinrainbow/json-schema are using the backslash ("`\`") in
error on occasion.

This fixes the issue in **UriResolver** as well as in **UriRetriever**.

With this commit, the *overall* usage of that DIRECTORY_SEPARATOR
constant has been redacted out of this project.

Additionally in a refactoring duplicate code has been removed from the
classes afterwards and some minor CS fixes have been applied.