Swap values in tests (in violation of spec)
This commit deliberately tests in a way that violates the spec. Versions
of this library prior to 6.0.0 allow properties adjacent to $ref to
override properties with the same name from the $ref'ed schema. This is
explicitly in violation of the spec, but cannot be changed in 5.x.x, as
it results in a significant behavioral change that may break end users'
application logic.

This issue is resolved in 6.0.0, and siblings of $ref are correctly
ignored from 6.0.0 onwards.