Remove large array type-specific performance optimisation
Fixes #441.

The bug in #441 was caused by refactoring of the optimisation which
introduced a type-checking error. Noting the performance impact is
negligible for all cases other than large arrays of strings or numbers,
and introduces significant cognitive complexity to a project that is
extremely short of maintainers, removing it seems like the best course
of action.

The performance improvement provided by this optimisation was
approximately 40%, however it also carried a number of other problematic
bugs - if it were to be reintroduced at a later date with those bugs
fixed (mainly the skipping of much of the validation logic for optimised
items, even in cases where that logic might be necessary), it would not
have such a significant impact.