Fix PHP 8 deprecation notices on required parameters appearing after optional parameters.
This is related to #619, where the fix was applied on the master branch. This commit backports them to 5.x.x branch.
From PHP 8.0 and forwards, a deprecation notice is raised if there are required parameters after optional parameters in a function/method signature.
See https://php.watch/versions/8.0/deprecate-required-param-after-optional for details.