Prevents infinite recursion
I've come across a case where it was possible to generate a recursive schema. In that instance this method goes into an infinite recursion.
I'm not 100% sure whether this is an issue with json-schema or with the fr3d/swagger-assertions layer that I'm using on top of it, but in any case adding an anti-infinite-recursion check here seems reasonable to me.