Use parallel-lint and cs2pr for improved feedback on linting errors (#812)
This pull request includes changes to the `.github/workflows/lint.yml`
file to improve the linting process for the project. The most important
changes include renaming the job from "tests" to "lint", updating the
PHP setup configuration, and modifying the linting command.

Improvements to linting workflow:

*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L12-R12):
Renamed the job from "tests" to "lint" to better reflect its purpose.
*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L30-R38):
Updated the PHP setup configuration to include `parallel-lint` and
`cs2pr` tools, and reordered the `php-version` and `coverage` parameters
for clarity.
*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L30-R38):
Modified the linting command to use `parallel-lint` with `cs2pr` for
better parallel processing and reporting.