feat: update to latest json schema test suite (#821)
This pull request updates the `json-schema/json-schema-test-suite`
dependency to a newer version and introduces a new test suite for
validating JSON schemas. The goal here is to have a quicker turnaround
then just the bowtie reports. The most important changes are grouped
below:

### Dependency Updates:
* Updated `json-schema/json-schema-test-suite` dependency in
`composer.json` from version `1.2.0` to `^23.2` and updated the package
version and reference in the `repositories` section to match the new
version.
[[1]](diffhunk://#diff-d2ab9925cad7eac58e0ff4cc0d251a937ecf49e4b6bf57f8b95aab76648a9d34L36-R36)
[[2]](diffhunk://#diff-d2ab9925cad7eac58e0ff4cc0d251a937ecf49e4b6bf57f8b95aab76648a9d34L62-R66)

### New Test Suite:
* Added a new `JsonSchemaTestSuite` class in
`tests/JsonSchemaTestSuite.php` to run validation tests using the
`json-schema/json-schema-test-suite`. This includes:
  - A `testIt` method to validate test cases and assert results.
- A `casesDataProvider` method to dynamically load test cases from the
`json-schema-test-suite` repository, skipping certain drafts.
- A `loadRemotesIntoStorage` helper method to load remote schemas into
the schema storage.