Update php-enum dependency version to ^4.4
Hi,

Versions of `mark-mabe/php-enum` lower than 4.4 emit a warning on PHP 8:

```
PHP Warning: Private methods cannot be final as they are never overridden by other classes
```

so it would be nice to restrict installs to versions that work fine. There is no BC breaks with this and no change in dependencies.